Initial structure Blocks with Factory pattern
A blueprint for FaBriCAD consists of building blocks. Any supported Building Block should inherit from AbstractBuildingBlock.

An AbstractBuildingBlock has a function render() that should generate the shapes for that blueprint. The function should always return a single Shape. Notice that a Container is a shape as well. To generate BuildingBlocks, we apply the Factory pattern.